refactor: extract RoomGridCell from RoomGrid (928→548 lines)
Move room cell template, computed helpers, and styling into dedicated
RoomGridCell.vue component. RoomGrid.vue now imports and delegates
to RoomGridCell, reducing its size by ~380 lines (AGENTS.md 250 LOC
limit compliance).
